Debug parser; parse concrete sets
The parser works! I also implemented parsing for concrete sets. The code is very
ugly, but it works, and it's fast. The segfaulting before was just due to the
fact that I didn't initialize _theorem to null in the TheoremProver -- it wasn't
a but in the parser.

I also removed a.out, which should never have been committed.