Add .gitignore and build script; start tests
I've organized the whole project structure now. The source files are in src/,
the tests are in tests/, and the binaries are in dist/. The build.sh file will
automatically compile all the source files. It can also build the testing
binary, and you can give it an option to build a version of the program for
debugging. I'm going to use Catch for testing.