Eliminate warnings; use pointer for quantified var
I got rid of some of the more obnoxious warnings by disabling them in the build
script, and I changed the code to fix the rest of them. The most frequent
warning involved copy constructors, which was all due to the edge case of
storing Symbols directly in the Quantified object, so I changed that to use a
pointer like everything else, even if it isn't strictly necessary. Now I don't
have to remember to use '.' instead of '->' as a special case, and I don't have
to worry about copy constructors.