Implement TheoremProver::Node
I have a much better idea of how TheoremProver is going to work than I used to.
It was a good thing that I used accessors for the _theorem, because there
actually isn't a _theorem member anymore -- it's just the root of the Node tree
now. The Node class started out as a POD struct but now it is quite a large
class. In particular, I implemented a pretty-printing method for it, which was
quite a bit of work.

The TheoremProver will work like this: when you set the theorem, it will make
that the root of its tree. It will then perform a depth-first pre-order
traversal, building the tree as it goes. The _stack will be used for this
traversal (it can't be done recursively because it doens't happen all at once;
the user steps it through). Once the whole tree has been traversed, the theorem
can be considered proven.

The givens of a particular node include the given in the node itself, if any,
plus all the givens of the parents all the way to the root. Instead of adding
parent pointers to the Node class (making it even more complex), I decided to
add a stack of Sentences to TheoremProver. These will get pushed and popped as
we traverse the tree so that it always contains the correct givens (i.e., things
that can be considered as facts in the proof of this subgoal). Since some nodes
have null givens, I will probably do a check before pushing and popping, rather
than wasting space with null givens.