Improve Node, rename stuff, refactor spa.cpp
Before TheoremProver was using the internals of Node directly. Now Node has more
public elements so that TP can interact with it on a higher level. Also, I
renamed some stuff, and I changed the way commands are parsed in spa.cpp so that
the same error message is printed whenever one of the 7 commands that require a
loaded theorem are entered, and there is no theorem.