Generalize relation negations
Before I had a _want on Sentence for general negation, but I ended up taking it
off. Now I've realized that it really makes to have it, but only on Relation.
Now I only define =, <, and <=, and treat !=, >=, and >, as their respective
negations. There is a parallel thing going on with (proper) subsets and set
(in)equality. I updated the parser accordingly. Now it's time to work on
decomposition and deduction.