Started implementing decomposition & deduction
It's going fairly well. There are a few little decisions to make. Some relate to
the balance between making things a pain for the user, and making everything
uniform. In general, decompositions should skip through the AND stage if
possible (go straight to two branches, not to a single AND first). Other things
could potentially go further but shouldn't (IFF just splits up into its two
implications, and then each can be proved directly or by contrapostive -- I
think this is fine).

There is going to be trouble with the quantifiers. Right now universal and
existential instantiation are the same. I will need to maintain a list of
variables or something to make them different.

I had another good idea -- instead of typing "triv" a lot (trivial), the user
could type a sentence of justification, which gets placed in the final proof. I
think taht would be pretty cool.