Implement 'stat' command, add the rest of the cmds
I've implemented a 'stat' comamnd which shows the theorem, the current goal, the
givens, and the number of goals left to prove. I've also fleshed out the user
interaface by adding the rest of the commands (I think). These are 'dec'
(decompose), 'ded' (deduce), 'triv' (prove a trivial goal), and 'just' (justify
a goal in words). SPA should be usable very soon.