Implement 'dec' and 'triv' (and more colours)
I made Decomp and Deduct classes, originally becuase I wanted to add a
destructor. Then I realized I couldn't do that because everything in the vectors
would get deleted, so I added a free method instead. I still think of them as
POD types (they have public fields), so maybe I shouldn't have done that. The
sentence.(h/c)pp files are getting quite large, so it might be time to refactor
and move stuff out.

There were a bunch of bugs with the tree pretty printing (which I hadn't tested
before, so that was expected). I fixed them all. I also made the current goal
label always print in red, just to emphasize it. I added a method to update the
_lineage of the TP, which works as long as it is called after each change to the
"current" node. It actually depends on the way we create the tree (specifically,
that the new current node will always be the left or right child of some node
already in the lineage). I implemented it by deleting everything in the lineage
until that parent node is found, and then adding the current node.

Finally, I implemented 'dec' (decomposition) and 'triv' (trivial proof). It
actually works now! It will navigate through the tree perfectly. Some of the
code in prover.cpp is starting to look ugly ... but I think it's OK.

Version 1.0 is on the horizon!